requirements.txt
attr==0.3.2
colorama==0.4.6
ConfigParser==7.1.0
cryptography==42.0.5
docutils==0.21.2
filelock==3.16.1
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.2
jnius==1.1.0
keyring==25.4.1
langchain==0.3.3
Pillow==10.4.0
protobuf==5.28.2
pyOpenSSL==24.2.1
pyOpenSSL==24.2.1
python-dotenv==1.0.1
redis==5.1.1
Sphinx==8.0.2
streamlit==1.36.0
thread==2.0.5
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
